+++ src/locale/en.json
@@ -56,6 +56,18 @@
                         "description": "My research relies on modelled environmental covariates sourced from <bLink>Brickman et al. 2021</bLink>, which features 1/12°, high-resolution projections of key covariates across the Northern Atlantic. This dataset has plenty of potential beyond my own work for use in other ecological forecast scenarios. This walkthrough presents a set of prepackaged functions and clear explanations for how anyone with a presence/absence dataset can build their own high resolution forecasts using the Brickman covariates."
                     }
                 }
+            },
+            "colobus": {
+                "miniTitle": "Emerging Work, ACM Animal-Computer Interaction (ACI) 2025",
+                "mainTitle": "Colobus Curio Cabinet: A Modular, Mirror-Based, Co-Designed Enrichment Proposal for <i>Colobus guereza</i>",
+                "subTitle": "Omi H. Johnson, Victoria Klimkowski, Shannon Haley, Cailigh MacDonald, Sarah Woodruff, Rebecca Kleinberger",
+                "description": 
+                "<par> In the modern age, animals are increasingly interacting with technology: service dogs must navigate automatic doors, smart collars can inform animal health, and zoo animals can interact with touch screens. Animal-Computer Interaction is a recently emerged field of research that studies the intersection of animals and technology. A key tenant of this field is that technology should be created using the principles of user-centered design and human-computer interaction, thus enabling natural and consented use of technology by animals. </par> <par> In this project within the <ialLink> Interact Animal Lab </ialLink>, we are constructing a sensor-enabled, screen- and mirror-based enrichment device for <i>Colobus Guereza</i> monkeys. The device consists of modular mirror components covered by movable flaps. The device is intended to encourage natural behaviors in <i>C. Guereza</i>, such as foraging and exploration of novel enrichment. Sensors and touch-screen modules would be used to collect data on effectiveness and add extra depth to play. This project is conducted in collaboration with a local zoo and offers critical insights on effective communication between zookeepers and researchers as well as best non-food enrichment strategies for primates. </par>",
+                "abstract": "This work explores complex non-food based enrichment for zoo-housed primates through the design, iterative co-design process, and proposed evaluation of a modular, mirror-based enrichment system developed for zoo-housed <i>Colobus guereza monkeys</i>. The interactive enrichment proposal, which leverages the Colobus troop's interest in reflective surfaces, proposes modular inserts containing various mirrors and screen-based interfaces within a wall-mounted \"Curio Cabinet\" with natural-looking module covers. The design of the Curio Cabinet aims to simulate natural exploration and problem-solving behaviors while promoting physical, cognitive, and sensory engagement. Drawing from principles in Animal-Computer Interaction (ACI) and animal enrichment literature, the proposal was iteratively co-designed with zoo professionals. This design contributes to the broader application of animal-centered design in zoo settings toward a new paradigm that supports animal's choice and control in their environmental interactions. The design also challenges conventional \"mirror-test\" approaches by exploring an animal's agency-driven, personal sense-making journey with reflective surfaces. The proposed deployment protocol opens opportunities to investigate new research questions around mirror enrichment and long-term engagement. Finally, the work provides a detailed account of a collaborative design pipeline and recommendations for future iterative co-design processes in complex zoo enrichment.",
+                "links": [
+                    "Watch the video presentation given at ACI 2025.",
+                    "Read the paper for free!"
+                ]
             }
         },
         "contact": {
